ppx: add [@drop_default] for record fields
For now it only works for record fields annotated with `[@option]`, by
dropping the field from JSON repr when the record value is `None`.

What's missing is to also support `[@drop_default]` for fields annotated
by `[@default X]` but we need to decide how to check for equality
between the default value and the field value.

One nice idea I had is to generate code like this:

```ocaml
type t = { a : int [@default 0] [@drop_default] }
...
let bnds =
  match [%equal int] a 0 with
  | true -> bnds
  | false -> ("a", a)::bnds
in
...
```

but this means this ppx will depends on another ppx which provides
`[%equal t]` deriver but sadly `ppx_compare` doesn't work with melange
now.