ppx: consistent errors in runtime
The main thing we fix here is the consistent use of exceptions in native
runtime, which used, before this commit either its own Of_json_error or
yojson's Type_error.

Then there's
```
val of_string : string -> t
```
which is not required by ppx but was introduced so that the runtime
modules can function as minimal JSON interfaces usable both in native
and browser environments (so called universal workflow).

Before this commit the error behaviour was not specified, now we
introduce `exception Of_json of string` in both native and browser and
make `of_string` raise it in case of invalid JSON.